+++ sc-tools/sc-config/src/sc_options.cpp
@@ -35,10 +35,10 @@ bool ScOptions::Has(std::vector<std::string> const & options) const
     std::string const & option = ScOptions::Upstream(item);
 
     if (std::find(m_tokens.begin(), m_tokens.end(), option) != m_tokens.end())
-      return SC_TRUE;
+      return true;
   }
 
-  return SC_FALSE;
+  return false;
 }
 
 std::string ScOptions::Upstream(std::string const & option)