+++ sc-tools/sc-config/tests/units/test_config.cpp
@@ -26,7 +26,7 @@ TEST(ScConfig, ValidConfig)
   EXPECT_EQ(std::string(params.log_type), "Console");
   EXPECT_EQ(std::string(params.log_file), "");
   EXPECT_EQ(std::string(params.log_level), "Debug");
-  EXPECT_EQ(std::string(params.binaries), "sc-machine-test-repo");
+  EXPECT_EQ(std::string(params.storage), "sc-machine-test-repo");
   EXPECT_EQ(std::string(params.extensions), "bin/extensions");
 }
 
@@ -63,7 +63,7 @@ TEST(ScConfig, DeprecatedRepoPath)
   EXPECT_EQ(std::string(params.log_type), "Console");
   EXPECT_EQ(std::string(params.log_file), "");
   EXPECT_EQ(std::string(params.log_level), "Debug");
-  EXPECT_EQ(std::string(params.binaries), "sc-machine-test-repo");
+  EXPECT_EQ(std::string(params.storage), "sc-machine-test-repo");
   EXPECT_EQ(std::string(params.extensions), "bin/extensions");
 }
 
@@ -82,6 +82,6 @@ TEST(ScConfig, DeprecatedExtensionsPath)
   EXPECT_EQ(std::string(params.log_type), "Console");
   EXPECT_EQ(std::string(params.log_file), "");
   EXPECT_EQ(std::string(params.log_level), "Debug");
-  EXPECT_EQ(std::string(params.binaries), "sc-machine-test-repo");
+  EXPECT_EQ(std::string(params.storage), "sc-machine-test-repo");
   EXPECT_EQ(std::string(params.extensions), "bin/extensions");
 }